refactor: move non-specific XMB functions into utils.js
